prueba_NatDigital/requirements.txt
anyio==2.2.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.4
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
decorator==4.4.2
defusedxml==0.7.0
distlib==0.3.1
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
entrypoints==0.3
filelock==3.0.12
idna==2.10
imutils==0.5.4
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.18.0
Jinja2==2.11.3
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab==3.0.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.2
opencv-contrib-python==4.5.1.48
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
prometheus-client==0.9.0
prompt-toolkit==3.0.16
pycparser==2.20
Pygments==2.8.0
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
terminado==0.9.2
testpath==0.4.4
toml==0.10.1
tornado==6.1
traitlets==5.0.5
urllib3==1.26.3
virtualenv==20.0.30
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
